+++ src/index.ts
@@ -196,7 +196,7 @@ export class BigAmount {
 
     let [num, den] = [1, 1]; // result if coef === 1 (i.e. x === 2 ** exp)
     if (coef !== 1) {
-      // Approximate `coef` using Farey sequences (or just binary search)
+      // Approximate `coef` using Farey sequences
       let [lnum, lden] = [1, 1];
       let [unum, uden] = [2, 1];
       let mid = 1.5;