+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.1.2 - 2021-03-14
+
+### Added
+
+- Test case for `fromNumber()` that is specific to the implementation using
+  Farey sequences.
+
+### Maintenance
+
+- Refactored test code related to `fromNumber()`.
+- Updated dev dependencies.
+
 ## v1.1.1 - 2021-03-13
 
 ### Added
@@ -7,9 +19,9 @@
 - Test case for `fromNumber()`.
 - Benchmark script for `fromNumber()`.
 
-### Other
+### Maintenance
 
-- Update dev dependencies.
+- Updated dev dependencies.
 
 ## v1.1.0 - 2021-03-10